backend/pyproject.toml
[project]
name = "houses-backend"
version = "0.1.0"
description = "UK house price performance API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "boto3>=1.34.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that require DynamoDB Local in Docker (deselect with -m 'not integration')",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
